start work on new script
A lot has changed. I'm at the point where I can sort the dates
without having it in the filename. I just need to put it in the
index.md file and it'll be fine. This solves the problem of
calling genblog twice too, since this is being done on the md files